+++ packages/angular/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/angular [1.0.0-alpha.94](https://github.com/anolilab/lunora/compare/@lunora/angular@1.0.0-alpha.93...@lunora/angular@1.0.0-alpha.94) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.112
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.65
+
+## @lunora/angular [1.0.0-alpha.93](https://github.com/anolilab/lunora/compare/@lunora/angular@1.0.0-alpha.92...@lunora/angular@1.0.0-alpha.93) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.111
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.64
+
 ## @lunora/angular [1.0.0-alpha.92](https://github.com/anolilab/lunora/compare/@lunora/angular@1.0.0-alpha.91...@lunora/angular@1.0.0-alpha.92) (2026-09-13)