+++ packages/codegen/CHANGELOG.md
@@ -1,3 +1,32 @@
+## @lunora/codegen [1.0.0-alpha.191](https://github.com/anolilab/lunora/compare/@lunora/codegen@1.0.0-alpha.190...@lunora/codegen@1.0.0-alpha.191) (2026-09-19)
+
+### Features
+
+* **server:** authorize whisper topics with an onWhisper hook ([#755](https://github.com/anolilab/lunora/issues/755)) ([e72c1b6](https://github.com/anolilab/lunora/commit/e72c1b60d6cc82f14c8c0b892c6f92a0e3aba93c))
+
+
+### Dependencies
+
+* **@lunora/advisor:** upgraded to 1.0.0-alpha.140
+* **@lunora/agent:** upgraded to 1.0.0-alpha.118
+* **@lunora/do:** upgraded to 1.0.0-alpha.147
+* **@lunora/server:** upgraded to 1.0.0-alpha.131
+
+## @lunora/codegen [1.0.0-alpha.190](https://github.com/anolilab/lunora/compare/@lunora/codegen@1.0.0-alpha.189...@lunora/codegen@1.0.0-alpha.190) (2026-09-19)
+
+### Features
+
+* **server:** drop a stale patch instead of clobbering a newer edit ([#762](https://github.com/anolilab/lunora/issues/762)) ([7641eea](https://github.com/anolilab/lunora/commit/7641eea6a4e3e4ea7588baa0ba4479f03778599f))
+
+
+### Dependencies
+
+* **@lunora/advisor:** upgraded to 1.0.0-alpha.139
+* **@lunora/agent:** upgraded to 1.0.0-alpha.117
+* **@lunora/do:** upgraded to 1.0.0-alpha.146
+* **@lunora/server:** upgraded to 1.0.0-alpha.130
+* **@lunora/shard-engine:** upgraded to 1.0.0-alpha.73
+
 ## @lunora/codegen [1.0.0-alpha.189](https://github.com/anolilab/lunora/compare/@lunora/codegen@1.0.0-alpha.188...@lunora/codegen@1.0.0-alpha.189) (2026-09-13)
 
 ### ⚠ BREAKING CHANGES