+++ packages/nuxt/CHANGELOG.md
@@ -1,3 +1,18 @@
+## @lunora/nuxt [1.0.0-alpha.138](https://github.com/anolilab/lunora/compare/@lunora/nuxt@1.0.0-alpha.137...@lunora/nuxt@1.0.0-alpha.138) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.112
+
+## @lunora/nuxt [1.0.0-alpha.137](https://github.com/anolilab/lunora/compare/@lunora/nuxt@1.0.0-alpha.136...@lunora/nuxt@1.0.0-alpha.137) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.111
+* **@lunora/runtime:** upgraded to 1.0.0-alpha.126
+
 ## @lunora/nuxt [1.0.0-alpha.136](https://github.com/anolilab/lunora/compare/@lunora/nuxt@1.0.0-alpha.135...@lunora/nuxt@1.0.0-alpha.136) (2026-09-13)