+++ packages/seed/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/seed [1.0.0-alpha.135](https://github.com/anolilab/lunora/compare/@lunora/seed@1.0.0-alpha.134...@lunora/seed@1.0.0-alpha.135) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.131
+* **@lunora/testing:** upgraded to 1.0.0-alpha.176
+
+## @lunora/seed [1.0.0-alpha.134](https://github.com/anolilab/lunora/compare/@lunora/seed@1.0.0-alpha.133...@lunora/seed@1.0.0-alpha.134) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.130
+* **@lunora/testing:** upgraded to 1.0.0-alpha.175
+
 ## @lunora/seed [1.0.0-alpha.133](https://github.com/anolilab/lunora/compare/@lunora/seed@1.0.0-alpha.132...@lunora/seed@1.0.0-alpha.133) (2026-09-13)