+++ packages/vite/CHANGELOG.md
@@ -1,3 +1,21 @@
+## @lunora/vite [1.0.0-alpha.226](https://github.com/anolilab/lunora/compare/@lunora/vite@1.0.0-alpha.225...@lunora/vite@1.0.0-alpha.226) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.191
+* **@lunora/config:** upgraded to 1.0.0-alpha.226
+* **@lunora/studio:** upgraded to 1.0.0-alpha.189
+
+## @lunora/vite [1.0.0-alpha.225](https://github.com/anolilab/lunora/compare/@lunora/vite@1.0.0-alpha.224...@lunora/vite@1.0.0-alpha.225) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.190
+* **@lunora/config:** upgraded to 1.0.0-alpha.225
+* **@lunora/studio:** upgraded to 1.0.0-alpha.188
+
 ## @lunora/vite [1.0.0-alpha.224](https://github.com/anolilab/lunora/compare/@lunora/vite@1.0.0-alpha.223...@lunora/vite@1.0.0-alpha.224) (2026-09-13)